requirements.txt
firebase_admin==6.5.0
Flask==3.0.0
Flask_Cors==4.0.0
Flask_RESTful==0.3.10
joblib==1.3.2
langchain==0.2.15
langchain_core==0.2.36
langchain_google_firestore==0.3.0
langchain_openai==0.1.23
langgraph==0.2.14
numpy==1.24.4
pandas==2.2.2
python-dotenv==1.0.1
typing_extensions==4.12.2
google-auth==2.34.0
mysql-connector-python
python-dotenv
openai==1.55.3
